Add Japanese mode support and update package versions
The code modifications are mainly adding a Japanese mode, depending on which 'lang' (language) configuration is loaded. It then passes if the mode is Japanese or not to a few function calls. Additionally, numerous version changes for different packages in the 'poetry.lock' file, ensuring the application is using the most current and stable libraries.